requirements/requirements_windows.txt
absl-py==2.3.1
accelerate==1.8.1
aesedb==0.1.6
ailment==9.2.95
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioquic==1.2.0
aiosignal==1.4.0
aiosmb==0.4.11
aiowinreg==0.0.12
altgraph==0.17.4
amqp==5.2.0
angr==9.2.95
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appdirs==1.4.4
archinfo==9.2.95
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asciimatics==1.15.0
asgiref==3.5.2
asn1crypto==1.5.1
astroid==3.3.10
astunparse==1.6.3
asyauth==0.0.21
asysocks==0.2.13
attrs==25.3.0
autopep8==2.3.2
azure-core==1.35.0
azure-storage-blob==12.19.0
bcrypt==4.3.0
billiard==4.2.1
binwalk==2.1.0
bitarray==3.4.3
bitsandbytes==0.46.1
bitstring==4.3.1
black==25.1.0
blinker==1.9.0
boto3==1.38.41
botocore==1.38.46
Brotli==1.0.9
build==1.2.1
cachetools==6.1.0
cairocffi==1.7.1
capstone==5.0.6
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
claripy==9.2.95
cle==9.2.95
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
CppHeaderParser==2.7.4
cryptography==44.0.3
cssselect2==0.8.0
cupy-cuda12x==13.4.1
cycler==0.12.1
dacite==1.8.1
dask==2025.5.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distributed==2025.5.1
distro==1.9.0
dnspython==2.7.0
dpcpp-cpp-rt==2025.0.5
dpkt==1.9.8
email_validator==2.2.0
enum-compat==0.0.3
fastapi==0.111.0
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
first==2.0.2
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.5
frida==17.2.4
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.1.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpie==3.2.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
hyperframe==6.0.0
idna==3.8
importlib-metadata==6.8.0
intel-cmplr-lib-rt==2025.0.5
intel-cmplr-lib-ur==2025.0.5
intel-cmplr-lic-rt==2025.0.5
intel-opencl-rt==2025.0.5
intel-openmp==2025.0.5
intel-pti==0.10.1
intel-sycl-rt==2025.0.5
intel_extension_for_pytorch==2.7.10+xpu
intervaltree==3.1.0
isodate==0.7.2
isort==6.0.1
itanium-demangler==1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.6
jiter==0.6.1
jmespath==1.0.0
joblib==1.5.1
jsonpath-ng==1.6.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaitaistruct==0.10
keras==3.10.0
keyring==24.2.0
keystone-engine==0.9.2
kiwisolver==1.4.8
kombu==5.5.2
ldap3==2.9.1
libclang==18.1.1
lief==0.16.6
llama_cpp_python==0.3.11
llvmlite==0.44.0
locket==1.0.0
loguru==0.7.2
lxml==6.0.0
macholib==1.16.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
mdurl==0.1.2
minidump==0.0.24
minikerberos==0.4.6
mitmproxy==12.1.1
mitmproxy-windows==0.12.6
mitmproxy_rs==0.12.6
mkl==2025.0.1
mkl-dpcpp==2025.0.1
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.0.8
msldap==0.5.15
mulpyplexer==0.9
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
namex==0.1.0
nampa==0.1.1
networkx==3.1
nltk==3.9.1
numba==0.61.2
numpy==2.3.1
onemkl-sycl-blas==2025.0.1
onemkl-sycl-datafitting==2025.0.1
onemkl-sycl-dft==2025.0.1
onemkl-sycl-lapack==2025.0.1
onemkl-sycl-rng==2025.0.1
onemkl-sycl-sparse==2025.0.1
onemkl-sycl-stats==2025.0.1
onemkl-sycl-vm==2025.0.1
onnx==1.18.0
onnxruntime==1.22.0
openai==1.90.0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
oscrypto==1.3.0
overrides==7.7.0
packaging==25.0
pandas==2.2.2
paramiko==3.5.1
partd==1.4.2
passlib==1.7.4
pathspec==0.12.1
pdfkit==1.0.0
pefile==2024.8.26
peft==0.15.2
pillow==11.3.0
pipdeptree==2.26.1
platformdirs==4.3.8
plumbum==1.9.0
ply==3.11
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.19.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydivert==2.1.0
pydot==4.0.0
pydyf==0.11.0
pyelftools==0.32
pyfiglet==1.0.2
pyformlang==1.0.11
Pygments==2.19.2
pylint==3.3.7
pylsqpack==0.3.22
pymongo==4.13.2
PyNaCl==1.5.0
pynput==1.8.1
pyopencl==2025.2.4
pyOpenSSL==25.0.0
pyparsing==3.0.9
pyperclip==1.9.0
pyphen==0.17.2
pyproject_hooks==1.2.0
pypykatz==0.6.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
pyshark==0.6
PySMT==0.9.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fx==0.3.2
python-multipart==0.0.20
python-registry==1.3.1
pytools==2025.2.1
pytorch-triton-xpu==3.3.0
pytz==2025.2
pyusb==1.3.1
pyvex==9.2.95
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qiling==1.4.6
questionary==2.1.0
r2pipe==1.9.4
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rpds-py==0.26.0
rpyc==6.0.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.2
s3transfer==0.13.0
safetensors==0.5.3
scapy==2.6.1
scikit-learn==1.5.1
scipy==1.16.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
siphash24==1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
sympy==1.14.0
tabulate==0.9.0
tbb==2022.2.0
tblib==3.1.0
tcmlib==1.2.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
tomlkit==0.13.3
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5
tqdm==4.66.4
transformers==4.49.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
umf==0.9.1
unicodecsv==0.14.1
unicorn==2.0.1.post1
unicrypto==0.0.10
unique_log_filter==0.1.0
urllib3==2.5.0
urwid==2.6.15
uvicorn==0.35.0
vine==5.1.0
volatility3==2.26.0
vulture==2.14
wasm==1.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
wheel==0.43.0
win32_setctime==1.2.0
winacl==0.1.9
windows-curses==2.4.1
WMI==1.5.1
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
yamale==5.2.1
yara-python==4.5.4
yarl==1.20.1
z3-solver==4.10.2.0
zict==3.0.0
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
zopfli==0.2.3.post1
zstandard==0.19.0
